add some instruction prototype

+++ include/decoder.h
@@ -62,8 +62,8 @@ uint8_t GetRs1(const uint32_t instruction);
 uint8_t GetRs2(const uint32_t instruction);
 
 uint32_t GetImm(const uint32_t instruction, const RV64InstructionFormatType type);
-uint32_t GetShamt(const uint32_t instruction);
-uint32_t GetCsr(const uint32_t instruction);
+uint8_t  GetShamt(const uint32_t instruction);
+uint16_t GetCsr(const uint32_t instruction);
 
 } // namespace decoder
 } // namespace rv64_emulator
